+++ lib/merchant_sidekick/active_merchant/gateways/authorize_net_gateway.rb
@@ -1,4 +1,5 @@
-# Implements the authorize.net specific gateway configuration
+# Implements the AuthorizeNet specific active merchant gateway wrapper
+# to configure the gateway properly.
 module MerchantSidekick
   module ActiveMerchant
     module Gateways