+++ lib/merchant_sidekick/active_merchant/gateways/bogus_gateway.rb
@@ -1,0 +1,19 @@
+# Implements the bogus gatway active merchant gateway wrapper
+module MerchantSidekick
+  module ActiveMerchant
+    module Gateways
+      class BogusGateway < ::MerchantSidekick::ActiveMerchant::Gateways::Base
+        class << self
+
+          def gateway
+            unless @@gateway
+              @@gateway = ::ActiveMerchant::Billing::BogusGateway.new
+            end
+            @@gateway
+          end
+
+        end
+      end
+    end
+  end
+end