+++ spec/gateway_spec.rb
@@ -62,7 +62,7 @@ describe MerchantSidekick::ActiveMerchant::Gateways::AuthorizeNetGateway do
   end
 
   it "should return active merchant authorize_net gateway instance" do
-    MerchantSidekick::ActiveMerchant::Gateways::AuthorizeNetGateway.gateway.should be_instance_of(::ActiveMerchant::Billing::AuthorizeNetGateway)
+    2.times {MerchantSidekick::ActiveMerchant::Gateways::AuthorizeNetGateway.gateway.should be_instance_of(::ActiveMerchant::Billing::AuthorizeNetGateway)}
   end
 
   it "should return active merchant gateway instance for default_gateway type" do
@@ -94,13 +94,39 @@ describe MerchantSidekick::ActiveMerchant::Gateways::PaypalGateway do
   end
 
   it "should return active merchant paypal gateway instance" do
-    MerchantSidekick::ActiveMerchant::Gateways::PaypalGateway.gateway.should be_instance_of(::ActiveMerchant::Billing::PaypalGateway)
+    2.times {MerchantSidekick::ActiveMerchant::Gateways::PaypalGateway.gateway.should be_instance_of(::ActiveMerchant::Billing::PaypalGateway)}
   end
   
   it "should return active merchant gateway instance for default_gateway type" do
     MerchantSidekick::default_gateway = :paypal_gateway
     MerchantSidekick::default_gateway.should be_instance_of(::ActiveMerchant::Billing::PaypalGateway)
     MerchantSidekick::Gateway.default_gateway.should be_instance_of(::ActiveMerchant::Billing::PaypalGateway)
+  end
+
+  after(:all) do
+    MerchantSidekick::Gateway.default_gateway = ActiveMerchant::Billing::BogusGateway.new
+  end
+
+end
+
+
+describe MerchantSidekick::ActiveMerchant::Gateways::BogusGateway do
+
+  before(:each) do
+    MerchantSidekick::Gateway.default_gateway = nil
+    MerchantSidekick::Gateway.config          = nil
+    MerchantSidekick::Gateway.config_path     = nil
+    MerchantSidekick::Gateway.gateway         = nil
+  end
+
+  it "should return active merchant paypal gateway instance" do
+    2.times {MerchantSidekick::ActiveMerchant::Gateways::BogusGateway.gateway.should be_instance_of(::ActiveMerchant::Billing::BogusGateway)}
+  end
+  
+  it "should return active merchant gateway instance for default_gateway type" do
+    MerchantSidekick::default_gateway = :bogus_gateway
+    MerchantSidekick::default_gateway.should be_instance_of(::ActiveMerchant::Billing::BogusGateway)
+    MerchantSidekick::Gateway.default_gateway.should be_instance_of(::ActiveMerchant::Billing::BogusGateway)
   end
 
   after(:all) do