+++ spec/spec_helper.rb
@@ -250,5 +250,6 @@ module ActiveMerchant
 end
 
 ActiveMerchant::Billing::Base.mode = :test
-MerchantSidekick::Gateway.default_gateway = ActiveMerchant::Billing::BogusGateway.new
 ActiveMerchant::Billing::CreditCard.require_verification_value = true
+
+MerchantSidekick::default_gateway = :bogus_gateway